arm/dts: cw2217 vbatt-cutoff modify to 3340mv
As moto standard,the vbatt-cutoff should be 3400mv,
But the cw2217 fg's algorithm,it will retain up to 1% of capacity,
and if a case of long-term heavy load discharge, the shutdown voltage
will also be increased.

the current parameter of vbatt-cutoff is 3400mv, but
From the past test data. the minimun vbatt-cutoff is 3460.
so we have to discrease the vbatt-curoff voltage.

Change-Id: I5c28c50565d215cb5984f1976e139c13b9458f11
Signed-off-by: liuyc20 <liuyc20@lenovo.com>
Reviewed-on: https://gerrit.mot.com/2117156
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Huosheng Liao <liaohs@motorola.com>
Submit-Approved: Jira Key